@@ -13,7 +13,7 @@ extension Encodable{
     }
 }
 
-public enum FlickDirection: String, Codable {
+public enum FlickDirection: String, Codable, Sendable {
     case left
     case top
     case right
@@ -22,7 +22,7 @@ public enum FlickDirection: String, Codable {
 
 /// - 変換対象の言語を指定します。
 /// - specify language to convert
-public enum CustardLanguage: String, Codable {
+public enum CustardLanguage: String, Codable, Sendable {
     /// - 英語(アメリカ)に変換します
     /// - convert to American English
     case en_US
@@ -46,7 +46,7 @@ public enum CustardLanguage: String, Codable {
 
 /// - 入力方式を指定します。
 /// - specify input style
-public enum CustardInputStyle: String, Codable {
+public enum CustardInputStyle: String, Codable, Sendable {
     /// - 入力された文字をそのまま用います
     /// - use inputted characters directly
     case direct
@@ -58,11 +58,13 @@ public enum CustardInputStyle: String, Codable {
 
 /// - カスタードのバージョンを指定します。
 /// - specify custard version
-public enum CustardVersion: String, Codable {
+public enum CustardVersion: String, Codable, Sendable {
     case v1_0 = "1.0"
+    case v1_1 = "1.1"
+    case v1_2 = "1.2"
 }
 
-public struct CustardMetadata: Codable, Equatable {
+public struct CustardMetadata: Codable, Equatable, Sendable {
     public init(custard_version: CustardVersion, display_name: String) {
         self.custard_version = custard_version
         self.display_name = display_name
@@ -76,7 +78,7 @@ public struct CustardMetadata: Codable, Equatable {
     public var display_name: String
 }
 
-public struct Custard: Codable, Equatable {
+public struct Custard: Codable, Equatable, Sendable {
     public init(identifier: String, language: CustardLanguage, input_style: CustardInputStyle, metadata: CustardMetadata, interface: CustardInterface) {
         self.identifier = identifier
         self.language = language
@@ -116,7 +118,7 @@ extension Array where Element == Custard {
 
 /// - インターフェースのキーのスタイルです
 /// - style of keys
-public enum CustardInterfaceStyle: String, Codable {
+public enum CustardInterfaceStyle: String, Codable, Sendable {
     /// - フリック可能なキー
     /// - flickable keys
     case tenkeyStyle = "tenkey_style"
@@ -128,7 +130,7 @@ public enum CustardInterfaceStyle: String, Codable {
 
 /// - インターフェースのレイアウトのスタイルです
 /// - style of layout
-public enum CustardInterfaceLayout: Codable, Equatable {
+public enum CustardInterfaceLayout: Codable, Equatable, Sendable {
     /// - 画面いっぱいにマス目状で均等に配置されます
     /// - keys are evenly layouted in a grid pattern fitting to the screen
     case gridFit(CustardInterfaceLayoutGridValue)
@@ -149,7 +151,7 @@ public extension CustardInterfaceLayout{
         case grid_scroll
     }
 
-    func encode(to encoder: Encoder) throws {
+    func encode(to encoder: any Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
         switch self {
         case let .gridFit(value):
@@ -164,7 +166,7 @@ public extension CustardInterfaceLayout{
         }
     }
 
-    init(from decoder: Decoder) throws {
+    init(from decoder: any Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         let type = try container.decode(ValueType.self, forKey: .type)
         let rowCount = try container.decode(Double.self, forKey: .row_count)
@@ -179,7 +181,7 @@ public extension CustardInterfaceLayout{
     }
 }
 
-public struct CustardInterfaceLayoutGridValue: Equatable {
+public struct CustardInterfaceLayoutGridValue: Equatable, Sendable {
     public init(rowCount: Int, columnCount: Int) {
         self.rowCount = rowCount
         self.columnCount = columnCount
@@ -193,7 +195,7 @@ public struct CustardInterfaceLayoutGridValue: Equatable {
     public var columnCount: Int
 }
 
-public struct CustardInterfaceLayoutScrollValue: Equatable {
+public struct CustardInterfaceLayoutScrollValue: Equatable, Sendable {
     public init(direction: ScrollDirection, rowCount: Double, columnCount: Double) {
         self.direction = direction
         self.rowCount = rowCount
@@ -213,7 +215,7 @@ public struct CustardInterfaceLayoutScrollValue: Equatable {
     public var columnCount: Double
 
     /// - direction of scroll
-    public enum ScrollDirection: String, Codable{
+    public enum ScrollDirection: String, Codable, Sendable {
         case vertical
         case horizontal
     }
@@ -221,7 +223,7 @@ public struct CustardInterfaceLayoutScrollValue: Equatable {
 
 /// - 画面内でのキーの位置を決める指定子
 /// - the specifier of key's position in screen
-public enum CustardKeyPositionSpecifier: Hashable {
+public enum CustardKeyPositionSpecifier: Hashable, Sendable {
     /// - gridFitのレイアウトを利用した際のキーの位置指定子
     /// - position specifier when you use grid fit layout
     case gridFit(GridFitPositionSpecifier)
@@ -233,7 +235,7 @@ public enum CustardKeyPositionSpecifier: Hashable {
 
 /// - gridFitのレイアウトを利用した際のキーの位置指定子に与える値
 /// - values in position specifier when you use grid fit layout
-public struct GridFitPositionSpecifier: Codable, Hashable {
+public struct GridFitPositionSpecifier: Codable, Hashable, Sendable {
     public init(x: Int, y: Int, width: Int = 1, height: Int = 1) {
         self.x = x
         self.y = y
@@ -256,7 +258,7 @@ public struct GridFitPositionSpecifier: Codable, Hashable {
         case x, y, width, height
     }
 
-    public init(from decoder: Decoder) throws {
+    public init(from decoder: any Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         self.x = try container.decode(Int.self, forKey: .x)
         self.y = try container.decode(Int.self, forKey: .y)
@@ -268,7 +270,7 @@ public struct GridFitPositionSpecifier: Codable, Hashable {
 
 /// - gridScrollのレイアウトを利用した際のキーの位置指定子に与える値
 /// - values in position specifier when you use grid scroll layout
-public struct GridScrollPositionSpecifier: Codable, Hashable, ExpressibleByIntegerLiteral {
+public struct GridScrollPositionSpecifier: Codable, Hashable, ExpressibleByIntegerLiteral, Sendable {
     /// - 通し番号
     /// - index
     public var index: Int
@@ -290,7 +292,7 @@ public extension GridScrollPositionSpecifier {
 
 /// - インターフェース
 /// - interface
-public struct CustardInterface: Codable, Equatable {
+public struct CustardInterface: Codable, Equatable, Sendable {
     public init(keyStyle: CustardInterfaceStyle, keyLayout: CustardInterfaceLayout, keys: [CustardKeyPositionSpecifier : CustardInterfaceKey]) {
         self.keyStyle = keyStyle
         self.keyLayout = keyLayout
@@ -357,7 +359,7 @@ public extension CustardInterface {
             case key
         }
 
-        func encode(to encoder: Encoder) throws {
+        func encode(to encoder: any Encoder) throws {
             var container = encoder.container(keyedBy: CodingKeys.self)
             try container.encode(specifierType, forKey: .specifier_type)
             switch self.specifier{
@@ -373,7 +375,7 @@ public extension CustardInterface {
             }
         }
 
-        init(from decoder: Decoder) throws {
+        init(from decoder: any Decoder) throws {
             let container = try decoder.container(keyedBy: CodingKeys.self)
             let specifierType = try container.decode(SpecifierType.self, forKey: .specifier_type)
             switch specifierType{
@@ -397,7 +399,7 @@ public extension CustardInterface {
         }
     }
 
-    func encode(to encoder: Encoder) throws {
+    func encode(to encoder: any Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
         try container.encode(keyStyle, forKey: .key_style)
         try container.encode(keyLayout, forKey: .key_layout)
@@ -405,7 +407,7 @@ public extension CustardInterface {
         try container.encode(elements, forKey: .keys)
     }
 
-    init(from decoder: Decoder) throws {
+    init(from decoder: any Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         self.keyStyle = try container.decode(CustardInterfaceStyle.self, forKey: .key_style)
         self.keyLayout = try container.decode(CustardInterfaceLayout.self, forKey: .key_layout)
@@ -418,7 +420,7 @@ public extension CustardInterface {
 
 /// - キーのデザイン
 /// - design information of key
-public struct CustardKeyDesign: Codable, Equatable, Hashable {
+public struct CustardKeyDesign: Codable, Equatable, Hashable, Sendable {
     public init(label: CustardKeyLabelStyle, color: CustardKeyDesign.ColorType) {
         self.label = label
         self.color = color
@@ -427,7 +429,7 @@ public struct CustardKeyDesign: Codable, Equatable, Hashable {
     public var label: CustardKeyLabelStyle
     public var color: ColorType
 
-    public enum ColorType: String, Codable {
+    public enum ColorType: String, Codable, Sendable {
         case normal
         case special
         case selected
@@ -437,7 +439,7 @@ public struct CustardKeyDesign: Codable, Equatable, Hashable {
 
 /// - バリエーションのキーのデザイン
 /// - design information of key
-public struct CustardVariationKeyDesign: Codable, Equatable, Hashable {
+public struct CustardVariationKeyDesign: Codable, Equatable, Hashable, Sendable {
     public init(label: CustardKeyLabelStyle) {
         self.label = label
     }
@@ -447,36 +449,74 @@ public struct CustardVariationKeyDesign: Codable, Equatable, Hashable {
 
 /// - キーに指定するラベル
 /// - labels on the key
-public enum CustardKeyLabelStyle: Codable, Equatable, Hashable {
+public enum CustardKeyLabelStyle: Codable, Equatable, Hashable, Sendable {
     case text(String)
     case systemImage(String)
+    case mainAndSub(String, String)
 }
 
 public extension CustardKeyLabelStyle {
     private enum CodingKeys: CodingKey {
         case text
         case system_image
+        case type
+        case main
+        case sub
     }
 
-    private var key: CodingKeys {
-        switch self{
-        case .text: return .text
-        case .systemImage: return .system_image
-        }
+    private enum ValueType: String, Codable {
+        case text
+        case system_image
+        case main_and_sub
     }
 
-    func encode(to encoder: Encoder) throws {
+    func encode(to encoder: any Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
         switch self {
-        case let .text(value),
-             let .systemImage(value):
-            try container.encode(value, forKey: key)
+        case let .text(value):
+            try container.encode(value, forKey: .text)
+        case let .systemImage(value):
+            try container.encode(value, forKey: .system_image)
+        case let .mainAndSub(main, sub):
+            try container.encode(ValueType.main_and_sub, forKey: .type)
+            try container.encode(main, forKey: .main)
+            try container.encode(sub, forKey: .sub)
         }
     }
 
-    init(from decoder: Decoder) throws {
+    init(from decoder: any Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
-        guard let key = container.allKeys.first else{
+        // "type"が見つかった場合
+        if let type = try? container.decode(ValueType.self, forKey: .type) {
+            switch type {
+            case .text:
+                let value = try container.decode(
+                    String.self,
+                    forKey: .text
+                )
+                self = .text(value)
+            case .system_image:
+                let value = try container.decode(
+                    String.self,
+                    forKey: .system_image
+                )
+                self = .systemImage(value)
+            case .main_and_sub:
+                let main = try container.decode(
+                    String.self,
+                    forKey: .main
+                )
+                let sub = try container.decode(
+                    String.self,
+                    forKey: .sub
+                )
+                self = .mainAndSub(main, sub)
+            }
+            return
+        }
+
+        // それ以外の場合(old cases)
+        guard container.allKeys.count == 1, let key = container.allKeys.first else {
             throw DecodingError.dataCorrupted(
                 DecodingError.Context(
                     codingPath: container.codingPath,
@@ -497,18 +537,25 @@ public extension CustardKeyLabelStyle {
                 forKey: .system_image
             )
             self = .systemImage(value)
+        default:
+            throw DecodingError.dataCorrupted(
+                DecodingError.Context(
+                    codingPath: container.codingPath,
+                    debugDescription: "Unabled to decode CustardKeyLabelStyle."
+                )
+            )
         }
     }
 }
 
 /// - key's data in interface
-public enum CustardInterfaceKey: Equatable, Hashable {
+public enum CustardInterfaceKey: Equatable, Hashable, Sendable {
     case system(CustardInterfaceSystemKey)
     case custom(CustardInterfaceCustomKey)
 }
 
 /// - keys prepared in default
-public enum CustardInterfaceSystemKey: Codable, Equatable, Hashable {
+public enum CustardInterfaceSystemKey: Codable, Equatable, Hashable, Sendable {
     /// - the globe key
     case changeKeyboard
 
@@ -555,12 +602,12 @@ public extension CustardInterfaceSystemKey{
         }
     }
 
-    func encode(to encoder: Encoder) throws {
+    func encode(to encoder: any Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
         try container.encode(self.valueType, forKey: .type)
     }
 
-    init(from decoder: Decoder) throws {
+    init(from decoder: any Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         let type = try container.decode(ValueType.self, forKey: .type)
         switch type {
@@ -583,7 +630,7 @@ public extension CustardInterfaceSystemKey{
 }
 
 /// - keys you can defined
-public struct CustardInterfaceCustomKey: Codable, Equatable, Hashable {
+public struct CustardInterfaceCustomKey: Codable, Equatable, Hashable, Sendable {
     public init(design: CustardKeyDesign, press_actions: [CodableActionData], longpress_actions: CodableLongpressActionData, variations: [CustardInterfaceVariation]) {
         self.design = design
         self.press_actions = press_actions
@@ -661,7 +708,7 @@ public extension CustardInterfaceCustomKey {
         )
     }
 
-    struct SimpleInputArgument: Equatable, ExpressibleByStringLiteral {
+    struct SimpleInputArgument: Equatable, ExpressibleByStringLiteral, Sendable {
         public var label: String
         public var input: String
 
@@ -734,7 +781,7 @@ public extension CustardInterfaceCustomKey {
 }
 
 /// - variation of key, includes flick keys and selectable variations in pc style keyboard.
-public struct CustardInterfaceVariation: Codable, Equatable, Hashable {
+public struct CustardInterfaceVariation: Codable, Equatable, Hashable, Sendable {
     public init(type: VariationType, key: CustardInterfaceVariationKey) {
         self.type = type
         self.key = key
@@ -748,7 +795,7 @@ public struct CustardInterfaceVariation: Codable, Equatable, Hashable {
 
     /// - キーの変種の種類
     /// - type of key variation
-    public enum VariationType: Equatable, Hashable {
+    public enum VariationType: Equatable, Hashable, Sendable {
         /// - variation of flick
         /// - warning: when you use pc style, this type of variation would be ignored.
         case flickVariation(FlickDirection)
@@ -778,7 +825,7 @@ public extension CustardInterfaceVariation {
         }
     }
 
-    func encode(to encoder: Encoder) throws {
+    func encode(to encoder: any Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
         try container.encode(self.key, forKey: .key)
         try container.encode(self.valueType, forKey: .type)
@@ -790,7 +837,7 @@ public extension CustardInterfaceVariation {
         }
     }
 
-    init(from decoder: Decoder) throws {
+    init(from decoder: any Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         self.key = try container.decode(CustardInterfaceVariationKey.self, forKey: .key)
         let valueType = try container.decode(ValueType.self, forKey: .type)
@@ -805,7 +852,7 @@ public extension CustardInterfaceVariation {
 }
 
 /// - data of variation key
-public struct CustardInterfaceVariationKey: Codable, Equatable, Hashable {
+public struct CustardInterfaceVariationKey: Codable, Equatable, Hashable, Sendable {
     public init(design: CustardVariationKeyDesign, press_actions: [CodableActionData], longpress_actions: CodableLongpressActionData) {
         self.design = design
         self.press_actions = press_actions
@@ -823,14 +870,14 @@ public struct CustardInterfaceVariationKey: Codable, Equatable, Hashable {
 }
 
 /// - Tab specifier
-public enum TabData: Hashable{
+public enum TabData: Hashable, Sendable {
     /// - tabs prepared by default
     case system(SystemTab)
     /// - tabs made as custom tabs.
     case custom(String)
 
     /// - system tabs
-    public enum SystemTab: String, Codable, Hashable{
+    public enum SystemTab: String, Codable, Hashable, Sendable {
         ///japanese input tab. the layout and input style depends on user's setting
         case user_japanese
 
@@ -860,10 +907,31 @@ public enum TabData: Hashable{
 
         ///the last tab
         case last_tab
+
+        /// clipboard history tab
+        case clipboard_history_tab
+
+        /// emoji tab
+        case emoji_tab
+
+        public init(from decoder: any Decoder) throws {
+            let container = try decoder.singleValueContainer()
+            let rawValue = try container.decode(String.self)
+            var key = rawValue[...]
+            // For debug build compatibility
+            if key.hasPrefix("__") {
+                key = key.dropFirst(2)
+            }
+            if let value = Self(rawValue: String(key)) {
+                self = value
+                return
+            }
+            throw DecodingError.dataCorruptedError(in: container, debugDescription: "unknown system tab: \(rawValue)")
+        }
     }
 }
 
-public struct ScanItem: Hashable {
+public struct ScanItem: Hashable, Sendable {
     public init(targets: [String], direction: ScanItem.Direction) {
         self.targets = targets
         self.direction = direction
@@ -872,13 +940,13 @@ public struct ScanItem: Hashable {
     public var targets: [String]
     public var direction: Direction
 
-    public enum Direction: String, Codable {
+    public enum Direction: String, Codable, Sendable {
         case forward
         case backward
     }
 }
 
-public struct LaunchItem: Hashable {
+public struct LaunchItem: Hashable, Sendable {
     public init(scheme: LaunchableApplication, target: String) {
         self.scheme = scheme
         self.target = target
@@ -887,7 +955,7 @@ public struct LaunchItem: Hashable {
     public var scheme: LaunchableApplication
     public var target: String
 
-    public enum LaunchableApplication: String, Codable {
+    public enum LaunchableApplication: String, Codable, Sendable {
         case azooKey
         case shortcuts
     }
@@ -895,9 +963,13 @@ public struct LaunchItem: Hashable {
 
 /// - アクション
 /// - actions done in key pressing
-public enum CodableActionData: Codable, Hashable {
+public enum CodableActionData: Codable, Hashable, Sendable {
     /// - input action specified character
     case input(String)
+
+    /// - input action specified character
+    /// - note: WIP. This action can be removed at any time.
+    case paste
 
     /// - exchange character "あ→ぁ", "は→ば", "a→A"
     case replaceDefault
@@ -962,6 +1034,7 @@ public extension CodableActionData{
 
     private enum ValueType: String, Codable{
         case input
+        case paste
         case replace_default
         case replace_last_characters
         case delete
@@ -977,6 +1050,8 @@ public extension CodableActionData{
         case toggle_caps_lock_state
         case dismiss_keyboard
         case launch_application
+        // This case is left for debug build compatibility.
+        case __paste
     }
 
     private var key: ValueType {
@@ -997,6 +1072,7 @@ public extension CodableActionData{
         case .toggleCapsLockState: return .toggle_caps_lock_state
         case .toggleCursorBar: return .toggle_cursor_bar
         case .toggleTabBar: return .toggle_tab_bar
+        case .paste: return .paste
         }
     }
 
@@ -1037,7 +1113,7 @@ public extension CodableActionData{
         }
     }
 
-    func encode(to encoder: Encoder) throws {
+    func encode(to encoder: any Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
         try container.encode(self.key, forKey: .type)
         switch self {
@@ -1055,11 +1131,11 @@ public extension CodableActionData{
             try container.encode(value.target, forKey: .target)
         case let .moveTab(value):
             try CodableTabArgument(tab: value).containerEncode(container: &container)
-        case .dismissKeyboard, .enableResizingMode, .toggleTabBar, .toggleCursorBar, .toggleCapsLockState, .complete, .smartDeleteDefault, .replaceDefault: break
+        case .dismissKeyboard, .enableResizingMode, .toggleTabBar, .toggleCursorBar, .toggleCapsLockState, .complete, .smartDeleteDefault, .replaceDefault, .paste: break
         }
     }
 
-    init(from decoder: Decoder) throws {
+    init(from decoder: any Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         let valueType = try container.decode(ValueType.self, forKey: .type)
         switch valueType {
@@ -1106,12 +1182,14 @@ public extension CodableActionData{
             let scheme = try container.decode(LaunchItem.LaunchableApplication.self, forKey: .scheme_type)
             let target = try container.decode(String.self, forKey: .target)
             self = .launchApplication(.init(scheme: scheme, target: target))
+        case .__paste, .paste:
+            self = .paste
         }
     }
 }
 
 
-public struct CodableLongpressActionData: Codable, Equatable, Hashable {
+public struct CodableLongpressActionData: Codable, Equatable, Hashable, Sendable {
     public static let none = CodableLongpressActionData()
     public init(start: [CodableActionData] = [], repeat: [CodableActionData] = []) {
         self.start = start
